+++ invevent/__init__.py
@@ -1,0 +1,2 @@
+"""Invevent bot and web modules."""
+
