+++ invevent/web/__init__.py
@@ -1,0 +1,3 @@
+from .main import app
+
+__all__ = ["app"]
